scripts/selector/requirements.txt
# AI股票选择器依赖包
# 核心框架
pyqlib>=0.9.0
akshare>=1.9.0
pandas>=1.3.0
numpy>=1.21.0

# 机器学习
lightgbm>=3.2.0
xgboost>=1.6.0
scikit-learn>=1.0.0

# 深度学习（可选）
torch>=1.12.0
transformers>=4.20.0

# 优化求解
cvxpy>=1.2.0
scipy>=1.7.0

# 数据处理
tqdm>=4.60.0
joblib>=1.1.0

# 可视化（可选）
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 工具库
pathlib2>=2.3.0
loguru>=0.6.0
pyyaml>=6.0

# 金融数据（备选）
yfinance>=0.1.70
tushare>=1.2.60

# 安装说明：
# pip install -r requirements.txt
# 
# 可选安装（按需）：
# pip install pyqlib --upgrade
# pip install akshare --upgrade
# pip install lightgbm xgboost
# pip install torch torchvision
# pip install cvxpy